bdd10

+++ app/main.py
@@ -75,16 +75,13 @@ async def actualizar_Incidente(incidente_id: int, incidenteU: IncidenteU):
 
 @app.delete("/Incidentes/{incidente_id}")
 async def eliminar_incidente(incidente_id: int, session: Session = Depends(get_session)):
-    for idx, existing_incidente in enumerate(Incidentes):
-        if existing_incidente.id == incidente_id:
-            del Incidentes[idx]  # Elimina el incidente de la lista.
-            return {"mensaje": "Incidente eliminado"}
-    raise HTTPException(status_code=404, detail="Incidente no encontrado")
+    incidente = session.get(Incidente, incidente_id)
+    if not incidente:
+        raise HTTPException(status_code=404, detail="Incidente no encontrado")
 
-session.delete(incidente)
-session.commit()
-
-return {"mensaje": "Incidente eliminado"}
+    session.delete(incidente)
+    session.commit()
+    return {"mensaje": "Incidente eliminado"}
 
 # Ruta para buscar incidentes por su descripción.
